+++ terraform/aws/analytical-platform-data-production/cjs-dashboard-app/iam-roles.tf
@@ -2,11 +2,11 @@ module "cjs_dashboard_app_iam_role" {
   #checkov:skip=CKV_TF_1:Module is from Terraform registry
 
   source  = "terraform-aws-modules/iam/aws//modules/iam-github-oidc-role"
-  version = "5.37.0"
+  version = "5.37.1"
 
   name = "github-cjs-dashboard-app"
 
-  subjects = ["ministryofjustice/cjs_scorecard_exploratory_analysis:*"]
+  subjects = ["ministryofjustice/cjs-dashboard:*"]
 
   policies = {
     cjs_dashboard_app = module.cjs_dashboard_app_iam_policy.arn