+++ terraform/aws/analytical-platform-landing-production/ithc/iam-roles.tf
@@ -1,6 +1,13 @@
+resource "aws_iam_policy" "quicksight_user" {
+  provider    = aws.analytical-platform-data-production
+  name        = "QuickSightPenTesterPolicy"
+  path        = "/"
+  description = "Create Admin QS User"
+  policy      = data.aws_iam_policy_document.quicksight_user.json
+}
 module "analytical_platform_development_pen_tester" {
   source  = "terraform-aws-modules/iam/aws//modules/iam-assumable-role"
-  version = "5.35.0"
+  version = "5.37.1"
   providers = {
     aws = aws.analytical-platform-deveopment
   }
@@ -17,7 +24,7 @@ module "analytical_platform_development_pen_tester" {
 
 module "analytical_platform_production_pen_tester" {
   source  = "terraform-aws-modules/iam/aws//modules/iam-assumable-role"
-  version = "5.35.0"
+  version = "5.37.1"
   providers = {
     aws = aws.analytical-platform-production
   }
@@ -37,7 +44,7 @@ module "analytical_platform_production_pen_tester" {
 
 module "analytical_platform_data_production_pen_tester" {
   source  = "terraform-aws-modules/iam/aws//modules/iam-assumable-role"
-  version = "5.35.0"
+  version = "5.37.1"
   providers = {
     aws = aws.analytical-platform-data-production
   }
@@ -51,6 +58,7 @@ module "analytical_platform_data_production_pen_tester" {
   role_requires_mfa     = true
   force_detach_policies = true
   custom_role_policy_arns = [
-    "arn:aws:iam::aws:policy/SecurityAudit"
+    "arn:aws:iam::aws:policy/SecurityAudit",
+    aws_iam_policy.quicksight_user.arn
   ]
 }