+++ terraform/aws/analytical-platform-production/cluster/terraform.tf
@@ -14,11 +14,11 @@ terraform {
     }
     aws = {
       source  = "hashicorp/aws"
-      version = "5.39.1"
+      version = "5.40.0"
     }
     kubernetes = {
       source  = "hashicorp/kubernetes"
-      version = "2.26.0"
+      version = "2.27.0"
     }
     random = {
       source  = "hashicorp/random"