+++ terraform/github/terraform.tf
@@ -10,11 +10,11 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "5.39.0"
+      version = "5.40.0"
     }
     github = {
       source  = "integrations/github"
-      version = "6.0.1"
+      version = "6.1.0"
     }
   }
   required_version = "~> 1.5"